Add fix_imports.py, update requirements, adjust sleep and load in tests
A new python file, fix_imports.py, was added to make the system path aware of the parent folder. The requirements.txt was updated by adding the requests module needed for the project. In the run_locust_tests.py file, the sleep time was reduced from 15 to 10 to speed up the performance, and the load settings were adjusted for testing.